Use a clusterrole and clusterrolebinding to watch the Additional CAs ConfigMap
We allow changing the namespace and the name of the configmap to track the additional CAs.
The previous role and rolebinding were created on the hard-coded namespace openshift-image-registry, preventing a user from using a configmap in another namespace.
This commit switches to using a ClusterRole and CluterRoleBinding and removes the role and rolebinding related to config maps local to the namespace of the operator.