The PodPlacementController need to use SecurityContextConstraints
When the PodPlacementController tries to patch a pod that requires a SecurityContextConstraint different than the default one in OCP (restricted-v2), the patch was not be admitted because the service account it uses was not allowed to "use" securitycontextconstrains.
This commit adds this permission in the PodPlacementController's ClusterRoleBinding. Since it can now use SCC different than the default, we also force the security context constraint specs in the deployments of the operands.